Move types into own modules
to keep files short and clear.

No public API changes.

ScmpSyscall is excluded ATM to avoid merge conflicts.
