+++ libseccomp/src/lib.rs
@@ -52,620 +52,27 @@ pub mod error;
 #[cfg(any(libseccomp_v2_5, doc))]
 pub mod notify;
 
+mod action;
+mod arch;
+mod arg_compare;
+mod compare_op;
+mod filter_attr;
+mod version;
+
 use error::ErrorKind::*;
 use error::{Result, SeccompError};
 use libseccomp_sys::*;
-use std::convert::TryInto;
 use std::ffi::{CStr, CString};
 use std::fmt;
 use std::os::unix::io::AsRawFd;
 use std::ptr::NonNull;
 
-/// Represents the version information of the libseccomp library.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct ScmpVersion {
-    pub major: u32,
-    pub minor: u32,
-    pub micro: u32,
-}
-
-impl ScmpVersion {
-    /// Gets the version of the currently loaded libseccomp library.
-    ///
-    /// This function returns `ScmpVersion` that represents the currently
-    /// loaded libseccomp version.
-    ///
-    /// # Errors
-    ///
-    /// If this function encounters an issue while getting the version,
-    /// an error will be returned.
-    pub fn current() -> Result<Self> {
-        if let Some(version) = unsafe { seccomp_version().as_ref() } {
-            Ok(Self {
-                major: version.major,
-                minor: version.minor,
-                micro: version.micro,
-            })
-        } else {
-            Err(SeccompError::new(Common(
-                "Could not get seccomp version".to_string(),
-            )))
-        }
-    }
-}
-
-impl From<(u32, u32, u32)> for ScmpVersion {
-    /// Creates a `ScmpVersion` from the specified arbitrary version.
-    ///
-    /// # Arguments
-    ///
-    /// * `version` - A tuple that represents the version of the libseccomp library.  
-    /// The index 0, 1, and 2 represent `major`, `minor`, and `micro` respectively.
-    fn from(version: (u32, u32, u32)) -> Self {
-        Self {
-            major: version.0,
-            minor: version.1,
-            micro: version.2,
-        }
-    }
-}
-
-impl fmt::Display for ScmpVersion {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}.{}.{}", self.major, self.minor, self.micro)
-    }
-}
-
-/// Represents filter attributes.
-///
-/// You can set/get the attributes of a filter context with
-/// [`ScmpFilterContext::set_filter_attr`] and [`ScmpFilterContext::get_filter_attr`] methods.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-#[non_exhaustive]
-pub enum ScmpFilterAttr {
-    /// The default filter action as specified in the call to seccomp reset.
-    ActDefault,
-    /// The filter action taken when the loaded filter does not
-    /// match the architecture of the executing application.
-    ActBadArch,
-    /// A flag to specify if the NO_NEW_PRIVS functionality should
-    /// be enabled before loading the seccomp filter into the kernel.
-    CtlNnp,
-    /// A flag to specify if the kernel should attempt to
-    /// synchronize the filters across all threads on seccomp load.
-    CtlTsync,
-    /// A flag to specify if the libseccomp should allow filter rules
-    /// to be created for the -1 syscall.
-    ApiTskip,
-    /// A flag to specify if the kernel should log all filter
-    /// actions taken except for the [`ScmpAction::Allow`] action.
-    CtlLog,
-    /// A flag to disable Speculative Store Bypass mitigations for
-    /// this filter.
-    CtlSsb,
-    /// A flag to specify the optimization level of the seccomp
-    /// filter.
-    CtlOptimize,
-    /// A flag to specify if the libseccomp should pass system error
-    /// codes back to the caller instead of the default -ECANCELED.
-    ApiSysRawRc,
-}
-
-impl ScmpFilterAttr {
-    fn to_sys(self) -> scmp_filter_attr {
-        match self {
-            Self::ActDefault => scmp_filter_attr::SCMP_FLTATR_ACT_DEFAULT,
-            Self::ActBadArch => scmp_filter_attr::SCMP_FLTATR_ACT_BADARCH,
-            Self::CtlNnp => scmp_filter_attr::SCMP_FLTATR_CTL_NNP,
-            Self::CtlTsync => scmp_filter_attr::SCMP_FLTATR_CTL_TSYNC,
-            Self::ApiTskip => scmp_filter_attr::SCMP_FLTATR_API_TSKIP,
-            Self::CtlLog => scmp_filter_attr::SCMP_FLTATR_CTL_LOG,
-            Self::CtlSsb => scmp_filter_attr::SCMP_FLTATR_CTL_SSB,
-            Self::CtlOptimize => scmp_filter_attr::SCMP_FLTATR_CTL_OPTIMIZE,
-            Self::ApiSysRawRc => scmp_filter_attr::SCMP_FLTATR_API_SYSRAWRC,
-        }
-    }
-}
-
-impl std::str::FromStr for ScmpFilterAttr {
-    type Err = SeccompError;
-
-    /// Converts string seccomp filter attribute to `ScmpFilterAttr`.
-    ///
-    /// # Arguments
-    ///
-    /// * `attr` - A string filter attribute, e.g. `SCMP_FLTATR_*`.
-    ///
-    /// See the [seccomp_attr_set(3)] man page for details on valid filter attribute values.
-    ///
-    /// [seccomp_attr_set(3)]: https://www.man7.org/linux/man-pages/man3/seccomp_attr_set.3.html
-    ///
-    /// # Errors
-    ///
-    /// If an invalid filter attribute is specified, an error will be returned.
-    fn from_str(attr: &str) -> Result<Self> {
-        match attr {
-            "SCMP_FLTATR_ACT_DEFAULT" => Ok(Self::ActDefault),
-            "SCMP_FLTATR_ACT_BADARCH" => Ok(Self::ActBadArch),
-            "SCMP_FLTATR_CTL_NNP" => Ok(Self::CtlNnp),
-            "SCMP_FLTATR_CTL_TSYNC" => Ok(Self::CtlTsync),
-            "SCMP_FLTATR_API_TSKIP" => Ok(Self::ApiTskip),
-            "SCMP_FLTATR_CTL_LOG" => Ok(Self::CtlLog),
-            "SCMP_FLTATR_CTL_SSB" => Ok(Self::CtlSsb),
-            "SCMP_FLTATR_CTL_OPTIMIZE" => Ok(Self::CtlOptimize),
-            "SCMP_FLTATR_API_SYSRAWRC" => Ok(Self::ApiSysRawRc),
-            _ => Err(SeccompError::new(ParseError)),
-        }
-    }
-}
-
-/// Represents a comparison operator which can be used in a filter rule.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-#[non_exhaustive]
-pub enum ScmpCompareOp {
-    /// Not equal
-    NotEqual,
-    /// Less than
-    Less,
-    /// Less than or equal
-    LessOrEqual,
-    /// Equal
-    Equal,
-    /// Greater than or equal
-    GreaterEqual,
-    /// Greater than
-    Greater,
-    /// Masked equality
-    ///
-    /// This works like `Equal` with the exception that the syscall argument is
-    /// masked with `mask` via an bitwise AND (i.e. you can check specific bits in the
-    /// argument).
-    MaskedEqual(#[doc = "mask"] u64),
-}
-
-impl ScmpCompareOp {
-    const fn to_sys(self) -> scmp_compare {
-        match self {
-            Self::NotEqual => scmp_compare::SCMP_CMP_NE,
-            Self::Less => scmp_compare::SCMP_CMP_LT,
-            Self::LessOrEqual => scmp_compare::SCMP_CMP_LE,
-            Self::Equal => scmp_compare::SCMP_CMP_EQ,
-            Self::GreaterEqual => scmp_compare::SCMP_CMP_GE,
-            Self::Greater => scmp_compare::SCMP_CMP_GT,
-            Self::MaskedEqual(_) => scmp_compare::SCMP_CMP_MASKED_EQ,
-        }
-    }
-}
-
-impl std::str::FromStr for ScmpCompareOp {
-    type Err = SeccompError;
-
-    /// Converts string seccomp comparison operator to `ScmpCompareOp`.
-    ///
-    /// # Arguments
-    ///
-    /// * `cmp_op` - A string comparison operator, e.g. `SCMP_CMP_*`.
-    ///
-    /// See the [seccomp_rule_add(3)] man page for details on valid comparison operator values.
-    ///
-    /// [seccomp_rule_add(3)]: https://www.man7.org/linux/man-pages/man3/seccomp_rule_add.3.html
-    ///
-    /// # Errors
-    ///
-    /// If an invalid comparison operator is specified, an error will be returned.
-    fn from_str(cmp_op: &str) -> Result<Self> {
-        match cmp_op {
-            "SCMP_CMP_NE" => Ok(Self::NotEqual),
-            "SCMP_CMP_LT" => Ok(Self::Less),
-            "SCMP_CMP_LE" => Ok(Self::LessOrEqual),
-            "SCMP_CMP_EQ" => Ok(Self::Equal),
-            "SCMP_CMP_GE" => Ok(Self::GreaterEqual),
-            "SCMP_CMP_GT" => Ok(Self::Greater),
-            "SCMP_CMP_MASKED_EQ" => Ok(Self::MaskedEqual(u64::default())),
-            _ => Err(SeccompError::new(ParseError)),
-        }
-    }
-}
-
-/// Represents a rule in a libseccomp filter context.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-#[repr(transparent)]
-pub struct ScmpArgCompare(scmp_arg_cmp);
-
-impl ScmpArgCompare {
-    /// Creates and returns a new condition to attach to a filter rule.
-    ///
-    /// The rule will match if the comparison of argument `arg` (zero-indexed argument
-    /// of the syscall) with the value provided by `datum` using the compare operator
-    /// provided by `op` is true.
-    ///
-    /// You can use the [`scmp_cmp!`](crate::scmp_cmp) macro instead of this to create
-    /// `ScmpArgCompare` in a more elegant way.
-    ///
-    /// # Arguments
-    ///
-    /// * `arg` - The number of the argument
-    /// * `op` - A comparison operator
-    /// * `datum` - A value to compare to
-    #[must_use]
-    pub const fn new(arg: u32, op: ScmpCompareOp, datum: u64) -> Self {
-        if let ScmpCompareOp::MaskedEqual(mask) = op {
-            Self(scmp_arg_cmp {
-                arg,
-                op: op.to_sys(),
-                datum_a: mask,
-                datum_b: datum,
-            })
-        } else {
-            Self(scmp_arg_cmp {
-                arg,
-                op: op.to_sys(),
-                datum_a: datum,
-                datum_b: 0,
-            })
-        }
-    }
-}
-
-impl From<ScmpArgCompare> for scmp_arg_cmp {
-    fn from(v: ScmpArgCompare) -> scmp_arg_cmp {
-        v.0
-    }
-}
-
-impl From<&ScmpArgCompare> for scmp_arg_cmp {
-    fn from(v: &ScmpArgCompare) -> scmp_arg_cmp {
-        v.0
-    }
-}
-
-#[rustfmt::skip]
-#[doc(hidden)]
-#[macro_export]
-macro_rules! __private_scmp_cmp_arg {
-    (arg0) => { 0 };
-    (arg1) => { 1 };
-    (arg2) => { 2 };
-    (arg3) => { 3 };
-    (arg4) => { 4 };
-    (arg5) => { 5 };
-}
-
-/// A macro to create [`ScmpArgCompare`] in a more elegant way.
-///
-/// ```
-/// use libseccomp::{ScmpArgCompare, ScmpCompareOp, scmp_cmp};
-///
-/// assert_eq!(
-///     scmp_cmp!($arg0 != 123),
-///     ScmpArgCompare::new(0, ScmpCompareOp::NotEqual, 123),
-/// );
-/// assert_eq!(
-///     scmp_cmp!($arg1 < 123),
-///     ScmpArgCompare::new(1, ScmpCompareOp::Less, 123),
-/// );
-/// assert_eq!(
-///     scmp_cmp!($arg2 <= 123),
-///     ScmpArgCompare::new(2, ScmpCompareOp::LessOrEqual, 123),
-/// );
-/// assert_eq!(
-///     scmp_cmp!($arg3 == 123),
-///     ScmpArgCompare::new(3, ScmpCompareOp::Equal, 123),
-/// );
-/// assert_eq!(
-///     scmp_cmp!($arg4 >= 123),
-///     ScmpArgCompare::new(4, ScmpCompareOp::GreaterEqual, 123),
-/// );
-/// assert_eq!(
-///     scmp_cmp!($arg5 > 123),
-///     ScmpArgCompare::new(5, ScmpCompareOp::Greater, 123),
-/// );
-/// assert_eq!(
-///     scmp_cmp!($arg0 & 0x0f0 == 123),
-///     ScmpArgCompare::new(0, ScmpCompareOp::MaskedEqual(0x0f0), 123),
-/// );
-/// ```
-#[macro_export]
-macro_rules! scmp_cmp {
-    ($_:tt $arg:tt != $datum:expr) => {
-        $crate::ScmpArgCompare::new(
-            $crate::__private_scmp_cmp_arg!($arg),
-            $crate::ScmpCompareOp::NotEqual,
-            $datum,
-        )
-    };
-    ($_:tt $arg:tt < $datum:expr) => {
-        $crate::ScmpArgCompare::new(
-            $crate::__private_scmp_cmp_arg!($arg),
-            $crate::ScmpCompareOp::Less,
-            $datum,
-        )
-    };
-    ($_:tt $arg:tt <= $datum:expr) => {
-        $crate::ScmpArgCompare::new(
-            $crate::__private_scmp_cmp_arg!($arg),
-            $crate::ScmpCompareOp::LessOrEqual,
-            $datum,
-        )
-    };
-    ($_:tt $arg:tt == $datum:expr) => {
-        $crate::ScmpArgCompare::new(
-            $crate::__private_scmp_cmp_arg!($arg),
-            $crate::ScmpCompareOp::Equal,
-            $datum,
-        )
-    };
-    ($_:tt $arg:tt >= $datum:expr) => {
-        $crate::ScmpArgCompare::new(
-            $crate::__private_scmp_cmp_arg!($arg),
-            $crate::ScmpCompareOp::GreaterEqual,
-            $datum,
-        )
-    };
-    ($_:tt $arg:tt > $datum:expr) => {
-        $crate::ScmpArgCompare::new(
-            $crate::__private_scmp_cmp_arg!($arg),
-            $crate::ScmpCompareOp::Greater,
-            $datum,
-        )
-    };
-    ($_:tt $arg:tt & $mask:tt == $datum:expr) => {
-        $crate::ScmpArgCompare::new(
-            $crate::__private_scmp_cmp_arg!($arg),
-            $crate::ScmpCompareOp::MaskedEqual($mask),
-            $datum,
-        )
-    };
-}
-
-/// Represents an action to be taken on a filter rule match in the libseccomp.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-#[non_exhaustive]
-pub enum ScmpAction {
-    /// Kills the process.
-    KillProcess,
-    /// Kills the thread.
-    KillThread,
-    /// Throws a SIGSYS signal.
-    Trap,
-    /// Triggers a userspace notification.  
-    /// NOTE: This action is only usable when the libseccomp API level 6
-    /// or higher is supported.
-    Notify,
-    /// Returns the specified error code.  
-    /// NOTE: You can only use integers from 0 to `u16::MAX`.
-    Errno(i32),
-    /// Notifies a tracing process with the specified value.
-    Trace(u16),
-    /// Allows the syscall to be executed after the action has been logged.
-    Log,
-    /// Allows the syscall to be executed.
-    Allow,
-}
-
-impl ScmpAction {
-    fn to_sys(self) -> u32 {
-        match self {
-            Self::KillProcess => SCMP_ACT_KILL_PROCESS,
-            Self::KillThread => SCMP_ACT_KILL_THREAD,
-            Self::Trap => SCMP_ACT_TRAP,
-            Self::Notify => SCMP_ACT_NOTIFY,
-            Self::Errno(x) => SCMP_ACT_ERRNO(x as u16),
-            Self::Trace(x) => SCMP_ACT_TRACE(x),
-            Self::Log => SCMP_ACT_LOG,
-            Self::Allow => SCMP_ACT_ALLOW,
-        }
-    }
-
-    fn from_sys(val: u32) -> Result<Self> {
-        match val & SCMP_ACT_MASK {
-            SCMP_ACT_KILL_PROCESS => Ok(Self::KillProcess),
-            SCMP_ACT_KILL_THREAD => Ok(Self::KillThread),
-            SCMP_ACT_TRAP => Ok(Self::Trap),
-            SCMP_ACT_NOTIFY => Ok(Self::Notify),
-            SCMP_ACT_ERRNO_MASK => Ok(Self::Errno(val as u16 as i32)),
-            SCMP_ACT_TRACE_MASK => Ok(Self::Trace(val as u16)),
-            SCMP_ACT_LOG => Ok(Self::Log),
-            SCMP_ACT_ALLOW => Ok(Self::Allow),
-            _ => Err(SeccompError::new(ParseError)),
-        }
-    }
-
-    /// Converts string seccomp action to `ScmpAction`.
-    ///
-    /// # Arguments
-    ///
-    /// * `action` - A string action, e.g. `SCMP_ACT_*`.
-    ///
-    /// See the [seccomp_rule_add(3)] man page for details on valid action values.
-    ///
-    /// [seccomp_rule_add(3)]: https://www.man7.org/linux/man-pages/man3/seccomp_rule_add.3.html
-    ///
-    /// # Errors
-    ///
-    /// If an invalid action is specified or a value on `"SCMP_ACT_TRACE"` is not in the
-    /// range from 0 to `u16::MAX`, an error will be returned.
-    pub fn from_str(action: &str, val: Option<i32>) -> Result<Self> {
-        match action {
-            "SCMP_ACT_KILL_PROCESS" => Ok(Self::KillProcess),
-            "SCMP_ACT_KILL_THREAD" => Ok(Self::KillThread),
-            "SCMP_ACT_KILL" => Ok(Self::KillThread),
-            "SCMP_ACT_TRAP" => Ok(Self::Trap),
-            "SCMP_ACT_NOTIFY" => Ok(Self::Notify),
-            "SCMP_ACT_ERRNO" => match val {
-                Some(v) => Ok(Self::Errno(v)),
-                None => Err(SeccompError::new(ParseError)),
-            },
-            "SCMP_ACT_TRACE" => match val {
-                Some(v) => Ok(Self::Trace(v.try_into()?)),
-                None => Err(SeccompError::new(ParseError)),
-            },
-            "SCMP_ACT_LOG" => Ok(Self::Log),
-            "SCMP_ACT_ALLOW" => Ok(Self::Allow),
-            _ => Err(SeccompError::new(ParseError)),
-        }
-    }
-}
-
-/// Represents a CPU architecture.
-/// Seccomp can restrict syscalls on a per-architecture basis.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-#[non_exhaustive]
-pub enum ScmpArch {
-    /// The native architecture token
-    Native,
-    /// The x86 (32-bit) architecture token
-    X86,
-    /// The x86-64 (64-bit) architecture token
-    X8664,
-    /// The x32 (32-bit x86_64) architecture token
-    X32,
-    /// The ARM architecture token
-    Arm,
-    /// The AARCH64 architecture token
-    Aarch64,
-    /// The MIPS architecture token
-    Mips,
-    /// The MIPS (64-bit) architecture token
-    Mips64,
-    /// The MIPS64N32 architecture token
-    Mips64N32,
-    /// The MIPSEL architecture token
-    Mipsel,
-    /// The MIPSEL (64-bit) architecture token
-    Mipsel64,
-    /// The MIPSEL64N32 architecture token
-    Mipsel64N32,
-    /// The PowerPC architecture token
-    Ppc,
-    /// The PowerPC (64-bit) architecture token
-    Ppc64,
-    /// The PowerPC64LE architecture token
-    Ppc64Le,
-    /// The S390 architecture token
-    S390,
-    /// The S390X architecture token
-    S390X,
-    /// The PA-RISC hppa architecture token
-    Parisc,
-    /// The PA-RISC (64-bit) hppa architecture token
-    Parisc64,
-    /// The RISC-V architecture token
-    Riscv64,
-}
-
-impl ScmpArch {
-    fn to_sys(self) -> u32 {
-        match self {
-            Self::Native => SCMP_ARCH_NATIVE,
-            Self::X86 => SCMP_ARCH_X86,
-            Self::X8664 => SCMP_ARCH_X86_64,
-            Self::X32 => SCMP_ARCH_X32,
-            Self::Arm => SCMP_ARCH_ARM,
-            Self::Aarch64 => SCMP_ARCH_AARCH64,
-            Self::Mips => SCMP_ARCH_MIPS,
-            Self::Mips64 => SCMP_ARCH_MIPS64,
-            Self::Mips64N32 => SCMP_ARCH_MIPS64N32,
-            Self::Mipsel => SCMP_ARCH_MIPSEL,
-            Self::Mipsel64 => SCMP_ARCH_MIPSEL64,
-            Self::Mipsel64N32 => SCMP_ARCH_MIPSEL64N32,
-            Self::Ppc => SCMP_ARCH_PPC,
-            Self::Ppc64 => SCMP_ARCH_PPC64,
-            Self::Ppc64Le => SCMP_ARCH_PPC64LE,
-            Self::S390 => SCMP_ARCH_S390,
-            Self::S390X => SCMP_ARCH_S390X,
-            Self::Parisc => SCMP_ARCH_PARISC,
-            Self::Parisc64 => SCMP_ARCH_PARISC64,
-            Self::Riscv64 => SCMP_ARCH_RISCV64,
-        }
-    }
-
-    fn from_sys(arch: u32) -> Result<Self> {
-        match arch {
-            SCMP_ARCH_NATIVE => Ok(Self::Native),
-            SCMP_ARCH_X86 => Ok(Self::X86),
-            SCMP_ARCH_X86_64 => Ok(Self::X8664),
-            SCMP_ARCH_X32 => Ok(Self::X32),
-            SCMP_ARCH_ARM => Ok(Self::Arm),
-            SCMP_ARCH_AARCH64 => Ok(Self::Aarch64),
-            SCMP_ARCH_MIPS => Ok(Self::Mips),
-            SCMP_ARCH_MIPS64 => Ok(Self::Mips64),
-            SCMP_ARCH_MIPS64N32 => Ok(Self::Mips64N32),
-            SCMP_ARCH_MIPSEL => Ok(Self::Mipsel),
-            SCMP_ARCH_MIPSEL64 => Ok(Self::Mipsel64),
-            SCMP_ARCH_MIPSEL64N32 => Ok(Self::Mipsel64N32),
-            SCMP_ARCH_PPC => Ok(Self::Ppc),
-            SCMP_ARCH_PPC64 => Ok(Self::Ppc64),
-            SCMP_ARCH_PPC64LE => Ok(Self::Ppc64Le),
-            SCMP_ARCH_S390 => Ok(Self::S390),
-            SCMP_ARCH_S390X => Ok(Self::S390X),
-            SCMP_ARCH_PARISC => Ok(Self::Parisc),
-            SCMP_ARCH_PARISC64 => Ok(Self::Parisc64),
-            SCMP_ARCH_RISCV64 => Ok(Self::Riscv64),
-            _ => Err(SeccompError::new(ParseError)),
-        }
-    }
-
-    /// Returns the system's native architecture.
-    ///
-    /// # Errors
-    ///
-    /// If this function encounters an issue while getting the native architecture,
-    /// an error will be returned.
-    pub fn native() -> Result<Self> {
-        let ret = unsafe { seccomp_arch_native() };
-
-        match Self::from_sys(ret) {
-            Ok(v) => Ok(v),
-            Err(_) => Err(SeccompError::new(Common(
-                "Could not get native architecture".to_string(),
-            ))),
-        }
-    }
-}
-
-impl std::str::FromStr for ScmpArch {
-    type Err = SeccompError;
-
-    /// Converts string seccomp architecture to `ScmpArch`.
-    ///
-    /// # Arguments
-    ///
-    /// * `arch` - A string architecture, e.g. `SCMP_ARCH_*`.
-    ///
-    /// See the [seccomp_arch_add(3)] man page for details on valid architecture values.
-    ///
-    /// [seccomp_arch_add(3)]: https://www.man7.org/linux/man-pages/man3/seccomp_arch_add.3.html
-    ///
-    /// # Errors
-    ///
-    /// If an invalid architecture is specified, an error will be returned.
-    fn from_str(arch: &str) -> Result<Self> {
-        match arch {
-            "SCMP_ARCH_NATIVE" => Ok(Self::Native),
-            "SCMP_ARCH_X86" => Ok(Self::X86),
-            "SCMP_ARCH_X86_64" => Ok(Self::X8664),
-            "SCMP_ARCH_X32" => Ok(Self::X32),
-            "SCMP_ARCH_ARM" => Ok(Self::Arm),
-            "SCMP_ARCH_AARCH64" => Ok(Self::Aarch64),
-            "SCMP_ARCH_MIPS" => Ok(Self::Mips),
-            "SCMP_ARCH_MIPS64" => Ok(Self::Mips64),
-            "SCMP_ARCH_MIPS64N32" => Ok(Self::Mips64N32),
-            "SCMP_ARCH_MIPSEL" => Ok(Self::Mipsel),
-            "SCMP_ARCH_MIPSEL64" => Ok(Self::Mipsel64),
-            "SCMP_ARCH_MIPSEL64N32" => Ok(Self::Mipsel64N32),
-            "SCMP_ARCH_PPC" => Ok(Self::Ppc),
-            "SCMP_ARCH_PPC64" => Ok(Self::Ppc64),
-            "SCMP_ARCH_PPC64LE" => Ok(Self::Ppc64Le),
-            "SCMP_ARCH_S390" => Ok(Self::S390),
-            "SCMP_ARCH_S390X" => Ok(Self::S390X),
-            "SCMP_ARCH_PARISC" => Ok(Self::Parisc),
-            "SCMP_ARCH_PARISC64" => Ok(Self::Parisc64),
-            "SCMP_ARCH_RISCV64" => Ok(Self::Riscv64),
-            _ => Err(SeccompError::new(ParseError)),
-        }
-    }
-}
+pub use action::ScmpAction;
+pub use arch::ScmpArch;
+pub use arg_compare::ScmpArgCompare;
+pub use compare_op::ScmpCompareOp;
+pub use filter_attr::ScmpFilterAttr;
+pub use version::ScmpVersion;
 
 /// Represents a syscall number.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
@@ -1900,123 +1307,6 @@ fn cvt(ret: i32) -> Result<()> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use std::str::FromStr;
-
-    #[test]
-    fn test_parse_filter_attr() {
-        let test_data = [
-            ("SCMP_FLTATR_ACT_DEFAULT", ScmpFilterAttr::ActDefault),
-            ("SCMP_FLTATR_ACT_BADARCH", ScmpFilterAttr::ActBadArch),
-            ("SCMP_FLTATR_CTL_NNP", ScmpFilterAttr::CtlNnp),
-            ("SCMP_FLTATR_CTL_TSYNC", ScmpFilterAttr::CtlTsync),
-            ("SCMP_FLTATR_API_TSKIP", ScmpFilterAttr::ApiTskip),
-            ("SCMP_FLTATR_CTL_LOG", ScmpFilterAttr::CtlLog),
-            ("SCMP_FLTATR_CTL_SSB", ScmpFilterAttr::CtlSsb),
-            ("SCMP_FLTATR_CTL_OPTIMIZE", ScmpFilterAttr::CtlOptimize),
-            ("SCMP_FLTATR_API_SYSRAWRC", ScmpFilterAttr::ApiSysRawRc),
-        ];
-        for data in test_data {
-            assert_eq!(
-                ScmpFilterAttr::from_str(data.0).unwrap().to_sys(),
-                data.1.to_sys()
-            );
-        }
-        assert!(ScmpFilterAttr::from_str("SCMP_INVALID_FLAG").is_err());
-    }
-
-    #[test]
-    fn test_parse_compare_op() {
-        let test_data = [
-            ("SCMP_CMP_NE", ScmpCompareOp::NotEqual),
-            ("SCMP_CMP_LT", ScmpCompareOp::Less),
-            ("SCMP_CMP_LE", ScmpCompareOp::LessOrEqual),
-            ("SCMP_CMP_EQ", ScmpCompareOp::Equal),
-            ("SCMP_CMP_GE", ScmpCompareOp::GreaterEqual),
-            ("SCMP_CMP_GT", ScmpCompareOp::Greater),
-            (
-                "SCMP_CMP_MASKED_EQ",
-                ScmpCompareOp::MaskedEqual(u64::default()),
-            ),
-        ];
-
-        for data in test_data {
-            assert_eq!(
-                ScmpCompareOp::from_str(data.0).unwrap().to_sys(),
-                data.1.to_sys()
-            );
-        }
-        assert!(ScmpCompareOp::from_str("SCMP_INVALID_FLAG").is_err());
-    }
-
-    #[test]
-    fn test_parse_action() {
-        let test_data = [
-            ("SCMP_ACT_KILL_PROCESS", ScmpAction::KillProcess),
-            ("SCMP_ACT_KILL_THREAD", ScmpAction::KillThread),
-            ("SCMP_ACT_KILL", ScmpAction::KillThread),
-            ("SCMP_ACT_TRAP", ScmpAction::Trap),
-            ("SCMP_ACT_NOTIFY", ScmpAction::Notify),
-            ("SCMP_ACT_ERRNO", ScmpAction::Errno(10)),
-            ("SCMP_ACT_TRACE", ScmpAction::Trace(10)),
-            ("SCMP_ACT_LOG", ScmpAction::Log),
-            ("SCMP_ACT_ALLOW", ScmpAction::Allow),
-        ];
-
-        for data in test_data {
-            if data.0 == "SCMP_ACT_ERRNO" || data.0 == "SCMP_ACT_TRACE" {
-                assert_eq!(
-                    ScmpAction::from_sys(ScmpAction::from_str(data.0, Some(10)).unwrap().to_sys())
-                        .unwrap(),
-                    data.1
-                );
-            } else {
-                assert_eq!(
-                    ScmpAction::from_sys(ScmpAction::from_str(data.0, None).unwrap().to_sys())
-                        .unwrap(),
-                    data.1
-                );
-            }
-        }
-        assert!(ScmpAction::from_str("SCMP_ACT_ERRNO", None).is_err());
-        assert!(ScmpAction::from_str("SCMP_ACT_TRACE", None).is_err());
-        assert!(ScmpAction::from_str("SCMP_INVALID_FLAG", None).is_err());
-        assert!(ScmpAction::from_sys(0x00010000).is_err());
-    }
-
-    #[test]
-    fn test_parse_arch() {
-        let test_data = [
-            ("SCMP_ARCH_NATIVE", ScmpArch::Native),
-            ("SCMP_ARCH_X86", ScmpArch::X86),
-            ("SCMP_ARCH_X86_64", ScmpArch::X8664),
-            ("SCMP_ARCH_X32", ScmpArch::X32),
-            ("SCMP_ARCH_ARM", ScmpArch::Arm),
-            ("SCMP_ARCH_AARCH64", ScmpArch::Aarch64),
-            ("SCMP_ARCH_MIPS", ScmpArch::Mips),
-            ("SCMP_ARCH_MIPS64", ScmpArch::Mips64),
-            ("SCMP_ARCH_MIPS64N32", ScmpArch::Mips64N32),
-            ("SCMP_ARCH_MIPSEL", ScmpArch::Mipsel),
-            ("SCMP_ARCH_MIPSEL64", ScmpArch::Mipsel64),
-            ("SCMP_ARCH_MIPSEL64N32", ScmpArch::Mipsel64N32),
-            ("SCMP_ARCH_PPC", ScmpArch::Ppc),
-            ("SCMP_ARCH_PPC64", ScmpArch::Ppc64),
-            ("SCMP_ARCH_PPC64LE", ScmpArch::Ppc64Le),
-            ("SCMP_ARCH_S390", ScmpArch::S390),
-            ("SCMP_ARCH_S390X", ScmpArch::S390X),
-            ("SCMP_ARCH_PARISC", ScmpArch::Parisc),
-            ("SCMP_ARCH_PARISC64", ScmpArch::Parisc64),
-            ("SCMP_ARCH_RISCV64", ScmpArch::Riscv64),
-        ];
-
-        for data in test_data {
-            assert_eq!(
-                ScmpArch::from_sys(ScmpArch::from_str(data.0).unwrap().to_sys()).unwrap(),
-                data.1
-            );
-        }
-        assert!(ScmpArch::from_str("SCMP_INVALID_FLAG").is_err());
-        assert!(ScmpArch::from_sys(1).is_err());
-    }
 
     #[test]
     fn test_ensure_supported_version() {
@@ -2028,46 +1318,6 @@ mod tests {
     fn test_ensure_supported_api() {
         assert!(ensure_supported_api("test", 3, ScmpVersion::from((2, 4, 0))).is_ok());
         assert!(ensure_supported_api("test", 100, ScmpVersion::from((2, 4, 0))).is_err());
-    }
-
-    #[test]
-    fn test_scmpargcompare() {
-        assert_eq!(
-            ScmpArgCompare::new(0, ScmpCompareOp::NotEqual, 8),
-            ScmpArgCompare(scmp_arg_cmp {
-                arg: 0,
-                op: scmp_compare::SCMP_CMP_NE,
-                datum_a: 8,
-                datum_b: 0,
-            })
-        );
-        assert_eq!(
-            ScmpArgCompare::new(0, ScmpCompareOp::MaskedEqual(0b0010), 2),
-            ScmpArgCompare(scmp_arg_cmp {
-                arg: 0,
-                op: scmp_compare::SCMP_CMP_MASKED_EQ,
-                datum_a: 0b0010,
-                datum_b: 2,
-            })
-        );
-        assert_eq!(
-            scmp_arg_cmp::from(ScmpArgCompare::new(0, ScmpCompareOp::NotEqual, 8)),
-            scmp_arg_cmp {
-                arg: 0,
-                op: scmp_compare::SCMP_CMP_NE,
-                datum_a: 8,
-                datum_b: 0,
-            }
-        );
-        assert_eq!(
-            scmp_arg_cmp::from(&ScmpArgCompare::new(0, ScmpCompareOp::NotEqual, 8)),
-            scmp_arg_cmp {
-                arg: 0,
-                op: scmp_compare::SCMP_CMP_NE,
-                datum_a: 8,
-                datum_b: 0,
-            }
-        );
     }
 
     #[test]
